Remove external dependency on nvidia-container-toolkit (i.e. nvidia-ctk)
Instead of relying on an external installation of the
nvidia-container-toolkit to provide us with the nvidia-ctk binary to use
for CDI hooks, we now bundle our own nvidia-ctk binary in the
k8s-dra-driver-gpu container image and install it on the host ourselves.

For now, the container image used to pull the nvidia-ctk image from is
hard-coded, but eventually we will want it to match the version of
whatever go dependency we have pulled in for the toolkit (to ensure
their features are in sync).

Note that it is still possible to pass a host path to the nvidia-ctk
binary (if desired), but this new feature makes it optional.

Signed-off-by: Kevin Klues <kklues@nvidia.com>