Allow ComputeDomain resources to be removed from any namespace
Prwviously, this sequence of events did not work properly:
1. Install the DRA driver in e.g. the `nvidia-dra-driver-gpu` namespace
2. Create a compute domain / run a workload
3. Uninstall the DRA driver
4. Reinstall the DRA driver in a new namespace, e.g. `gpu-operator`
5. Attempt to delete the workload and its corresponding ComputeDomain

When this happened, ComputeDomain deletion would hang on trying to delete
the `daemonset` because it was started in a different namespace.

To address this, this patch removes filtering by a specific namespace
for the daemonset and resourceclaimtemplate resources managed by the
compute domain controller.

Signed-off-by: Kevin Klues <kklues@nvidia.com>