Use device minor as suffix in device name instead of the index
The device minor is more stable than the index in terms of not changing
once a node has been brought online. The minor number may change across
a node reboot, but it wont change if one of the GPUs falls off the bus
or is explicitly drained / blocklisted from the driver (whereas the
index will adjust for any missing GPUs).

Signed-off-by: Kevin Klues <kklues@nvidia.com>