Add kubeVersion to helm chart
The kubeVersion puts constraints on the Kubernetes versions where this
helm chart can be installed. Note a '-0' at the end of a version string
in the kubeVersion generally indicates that all pre-release versions
should also be included. However, it is also requird to allow a given
chart to be installed on GKE as they add a '-gke-*' suffix to all of
their Kubernetes versions.

Signed-off-by: Kevin Klues <kklues@nvidia.com>