Conditionally mount the host's /dev over the container's /dev
Currently, in the compute domain driver, we unconditionally mount the
host's /dev over the container's /dev regardless of what the devRoot of
the GPU driver is. This change was introduced in
https://github.com/NVIDIA/k8s-dra-driver-gpu/pull/307.

Unfortunately, this causes some problems on certain systems as described
in https://github.com/NVIDIA/k8s-dra-driver-gpu/issues/477.

This patch makes this mount contingent on devRoot == "/" (which is the
only case where we actually ever wanted / needed to have this mount in
the first place.

Signed-off-by: Kevin Klues <kklues@nvidia.com>