Simplify logic in UpdateDNSNameMappings
Previously, we were always updating the ipToDNSName in place, but a
recent change was made to use a local copy instead. With that in place,
we no longer need to first remove old and then add new IPs to te map.
Instead we can just construct the new one only with the new IPs and
reassign it.

Signed-off-by: Kevin Klues <kklues@nvidia.com>