Do not swallow nvidia-imex-ctl stdout/stderr as part of `check`
Resolves #635.

Now, when nvidia-imex-ctl exits non-zero, the wrapper
(`compute-domain-daemon check`) logs the complete output
(stdout and stderr) of the child process.

That applies to startup probe, readiness probe, and liveness probe.

With that, a `kubectl describe ...` shows  for example:

    Startup probe failed: E1002 11:14:54.489874      71 main.go:353] \
       nvidia-imex-ctl failed (exit status 1), stdout/err: [...]

That may be verbose, but allows for faster debug cycles. Without this,
users have to for example shell into a container to run nvidia-imex-ctl
manually to get to this kind of output.

Signed-off-by: Dr. Jan-Philip Gehrcke <jgehrcke@nvidia.com>