build: retry TCP RST when curling bash source
Error seen:

curl: (7) Failed to connect to mirror.cs.odu.edu port 443 after 306 ms: Connection refused

By default, a TCP connection rejection (RST) is not treated
by curl as a transient error, see

https://curl.se/docs/manpage.html#--retry-connrefused

It's a transient error in the sense that it's often
a way to implement backpressure. We retry at slow rate.

`--retry-all-errors` is what we want here, it includes
`--retry-connrefused`.

Signed-off-by: Dr. Jan-Philip Gehrcke <jgehrcke@nvidia.com>